+++ packages/adminui-vue/components/controls/USelectMultiple.vue
@@ -175,7 +175,7 @@ const { debounce } = require('throttle-debounce')
 const clickOutsideDropdown = require('./mixins/clickOutsideDropdown')
 
 /**
- * When you need to select few values from entity use multiple select.
+ * Milty-select component mapped to Entity
  */
 export default {
   name: 'USelectMultiple',
@@ -184,7 +184,7 @@ export default {
 
   props: {
     /**
-     * Selected entity ID
+     * Selected IDs array
      * @model
      */
     value: {
@@ -230,7 +230,7 @@ export default {
      */
     readonly: Boolean,
     /**
-     * An array with IDs of elements that unable to remove
+     * An array with IDs of non-removable elements
      */
     fixedItems: {
       type: Array,
@@ -238,7 +238,7 @@ export default {
     },
 
     /**
-     * Search request condition
+     * Search by include (may be slow) or by first letters (faster)
      */
     searchStrategy: {
       type: String,
@@ -316,34 +316,20 @@ export default {
 
   watch: {
     /**
-     * when value changed need to check is item added or removed
-     * if added need to push formatted values (ID, label) to displayOptions
-     * if removed -> splice from displayOptions
+     * Update tags when value is changed
      */
     value: {
       immediate: true,
-      async handler (newVal, oldVal = []) {
-        const isAdded = newVal.length > oldVal.length
-        if (isAdded) {
-          if (!this.fixedItems.every(i => this.value.includes(i))) {
-            console.error('You should provide an initial value if you want items to be fixed')
-          }
-
-          const addedItems = newVal.filter(a => !oldVal.includes(a))
-          const formattedItems = await this.getFormattedOptions(addedItems) // temp
-          this.displayedOptions.push(...formattedItems)
-        } else {
-          const removedItems = oldVal.filter(a => !newVal.includes(a))
-          for (const item of removedItems) {
-            const index = this.displayedOptions.findIndex(o => o[this.valueAttribute] === item)
-            this.displayedOptions.splice(index, 1)
-          }
-        }
+      async handler (value) {
+        this.displayedOptions = await this.getFormattedOptions(value)
       }
     }
   },
 
   methods: {
+    /**
+     * @return {ClientRepository}
+     */
     getRepository () {
       if (this.repository) {
         return this.repository()
@@ -427,13 +413,12 @@ export default {
           })
         }
       }
-      const willFetched = result
-        .filter(o => !o.hasOwnProperty('label'))
+      const shouldFetch = result.filter(o => !Object.prototype.hasOwnProperty.call(o, 'label'))
         .map(o => o[this.valueAttribute])
 
-      if (willFetched.length > 0) {
-        const responseData = await this.fetchDisplayValues(willFetched)
-        for (const fetchedID of willFetched) {
+      if (shouldFetch.length) {
+        const responseData = await this.fetchDisplayValues(shouldFetch)
+        for (const fetchedID of shouldFetch) {
           const responseItem = responseData.find(i => i[this.valueAttribute] === fetchedID)
           const option = result.find(i => i[this.valueAttribute] === fetchedID)
           if (responseItem) {
@@ -667,146 +652,3 @@ export default {
   padding-right: 5px;
 }
 </style>
-
-<docs>
-One of these options is required:
-  - `entity-name`
-  - `repository`
-
-### Use as `entity-name`
-
-```vue
-<template>
-  <u-select-multiple
-    v-model="model"
-    entity-name="tst_dictionary"
-  />>
-</template>
-<script>
-  export default {
-    data () {
-      return {
-        value: []
-      }
-    }
-  }
-</script>
-```
-
-### Use as `repository`
-Need to set function which returns UB Repository
-
-```vue
-<template>
-  <u-select-multiple
-    v-model="model"
-    :repository="getRepo"
-  />
-</template>
-<script>
-  export default {
-    data () {
-      return {
-        value: []
-      }
-    },
-
-    methods: {
-      getRepo () {
-        return $UB.Repository('tst_maindata')
-          .attrs('ID', 'code', 'caption')
-          .where('parent', '=', 31231221312312) // TODO: set valid ID
-      }
-    }
-  }
-</script>
-```
-
-### Custom `valueAttribute`
-Need when you need to change default model propery.
-Its like attribute `value` in native `<option>` tag.
-For example when you need instead `ID` like `code`.
-
-```vue
-<template>
-  <u-select-multiple
-    v-model="model"
-    entity-name="tst_dictionary"
-    value-attribute="code"
-  />>
-</template>
-<script>
-  export default {
-    data () {
-      return {
-        value: []
-      }
-    }
-  }
-</script>
-```
-
-### Clearable
-
-```vue
-<template>
-  <u-select-multiple
-    v-model="model"
-    entity-name="tst_dictionary"
-    clearable
-  />
-</template>
-<script>
-  export default {
-    data () {
-      return {
-        value: [1,2,3]
-      }
-    }
-  }
-</script>
-```
-
-### Disabled
-
-```vue
-<template>
-  <u-select-multiple
-    v-model="model"
-    entity-name="tst_dictionary"
-    disabled
-  />
-</template>
-<script>
-  export default {
-    data () {
-      return {
-        value: [1,2,3]
-      }
-    }
-  }
-</script>
-```
-
-### filteredItems
-
-```vue
-<template>
-  <u-select-multiple
-    v-model="model"
-    entity-name="tst_dictionary"
-    :fixed-items="fixedItems"
-  />
-</template>
-<script>
-  export default {
-    data () {
-      return {
-        value: [1,2,3],
-        fixedItems: [2]
-      }
-    }
-  }
-</script>
-```
-</docs>