+++ packages/adminui-pub/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+ - `UB.core.UBUtil.getEntityAttributesTreeData` HTML for attribute label improved for better UX 
 
 ### Deprecated
 
@@ -15,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.20.6] - 2021-04-16
+## [5.20.5] - 2021-04-13
 ## [5.20.4] - 2021-04-02
 ## [5.20.3] - 2021-04-01
 ## [5.20.2] - 2021-03-30