+++ packages/openid-connect/CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.20.6] - 2021-04-16
+## [5.20.5] - 2021-04-13
 ## [5.20.4] - 2021-04-02
 ## [5.20.3] - 2021-04-01
 ## [5.20.2] - 2021-03-30