+++ packages/ubm/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+ - `ubm_navshortcut` form layout changed to `u-grid` + attributes tree UI improved
 
 ### Deprecated
 
@@ -15,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.20.6] - 2021-04-16
+## [5.20.5] - 2021-04-13
 ## [5.20.4] - 2021-04-02
 ### Fixed
  - ubm_query JSON to clob migration (for Oracle) not fails during migration from a version where  ubm_query table does not exists